requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.4
django-environ==0.4.4
django-rest-swagger==2.1.2
djangorestframework==3.8.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
requests==2.18.4
simplejson==3.14.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
factory_boy==2.11.1
Faker>=0.8.15
python-dateutil==2.7.3
text-unidecode==1.2

# factory boy
Faker==0.8.16
factory-boy==2.11.1
python-dateutil==2.7.3
text-unidecode==1.2